Separate out fmpz_poly test code and add fmpz_mod_poly test code. Document generic polynomials and fmpz_poly. Add documentation and tests for fmpz conversions. Move files and documentation sections around.

+++ src/flint/fmpz.jl
@@ -986,8 +986,8 @@ function convert(::Type{Int}, a::fmpz)
    return ccall((:fmpz_get_si, :libflint), Int, (Ptr{fmpz},), &a)
 end
 
-function convert(::Type{UInt64}, x::fmpz)
-   return ccall((:fmpz_get_ui, :libflint), UInt64, (Ptr{fmpz}, ), &x)
+function convert(::Type{UInt}, x::fmpz)
+   return ccall((:fmpz_get_ui, :libflint), UInt, (Ptr{fmpz}, ), &x)
 end
 
 function convert(::Type{Float64}, n::fmpz)
